Refactor: modularize Floor2D system - Split monolithic code into modules - Improved panning stability and performance
