+++ index.html
@@ -35,6 +35,7 @@
           <ul>
             <li><strong>Ctrl + Z:</strong> Undo last action</li>
             <li><strong>Alt:</strong> Move tool allow dragging vertex</li>
+            <li><strong>Alt + grab:</strong> navigate the canvas</li>
           </ul>
         </div>
       </div>
