+++ src/js/main.js
@@ -1,41 +1,52 @@
 import 'bootstrap/dist/js/bootstrap.bundle.min.js'
 import paper from 'paper'
-import { createGrid } from './2d/floor2d.js'
+import { createGrid, setup2D } from './2d/core2d.js'
 import { setupDxfUpload } from './2d/dxfLoader.js'
 import { init3D, camera, renderer } from './3d/scene3d.js'
 import { convertPathsTo3D } from './3d/pathsTo3d.js'
-// import parseDXF from 'dxf-parser'
 
-createGrid()
-setupDxfUpload()
-init3D()
+// ----------------------------------------------------
+// 1. Initialize 2D + 3D environments
+// ----------------------------------------------------
+setup2D()        // Sets up Paper.js, grid, and context menu
+setupDxfUpload() // Enables DXF import
+init3D()         // Sets up 3D scene
 
-// Switch between 2D and 3D modes
+// ----------------------------------------------------
+// 2. Switch between 2D and 3D modes
+// ----------------------------------------------------
 const container2D = document.getElementById('container2D')
 const container3D = document.getElementById('container3D')
 const switchButton = document.getElementById('switchMode')
+
 switchButton.addEventListener('click', () => {
   if (container3D.style.display === 'none') {
+    // Switch to 3D
     container2D.style.display = 'none'
     container3D.style.display = 'block'
     convertPathsTo3D()
     switchButton.textContent = 'Switch to 2D'
   } else {
+    // Switch to 2D
     container2D.style.display = 'block'
     container3D.style.display = 'none'
     switchButton.textContent = 'Switch to 3D'
-    // fixes when window is small that part of floorplan is cutted then turn to 3D the window is big again then turn to 2d the canvas is striched not resized
+
+    // Fix canvas stretch issue after resizing
     paper.view.viewSize = new paper.Size(window.innerWidth, window.innerHeight)
+    createGrid() // Recreate grid to fit new view bounds
   }
 })
 
-// Handle keyboard shortcuts toggle
+// ----------------------------------------------------
+// 3. Keyboard Shortcuts Toggle Panels
+// ----------------------------------------------------
 function setupShortcutsToggle(headerId, contentId) {
   const shortcutsHeader = document.getElementById(headerId)
   const shortcutsContent = document.getElementById(contentId)
   const shortcutsArrow = document.getElementsByClassName('shortcutsArrow')
 
-  let shortcutsExpanded = false // Start expanded
+  let shortcutsExpanded = false // Start collapsed
 
   shortcutsHeader.addEventListener('click', () => {
     shortcutsExpanded = !shortcutsExpanded
@@ -53,10 +64,10 @@ function setupShortcutsToggle(headerId, contentId) {
 setupShortcutsToggle('shortcutsHeader2D', 'shortcutsContent2D')
 setupShortcutsToggle('shortcutsHeader3D', 'shortcutsContent3D')
 
-// Handle window resizing
+// ----------------------------------------------------
+// 4. Handle window resizing (3D responsiveness)
+// ----------------------------------------------------
 window.addEventListener('resize', () => {
-  // console.log('resizing...')
-  // paper.view.viewSize = new paper.Size(window.innerWidth, window.innerHeight) # replaced by resize att in html
   camera.aspect = window.innerWidth / window.innerHeight
   camera.updateProjectionMatrix()
   renderer.setSize(window.innerWidth, window.innerHeight)
